Change the floated-left element to use float to allow text to wrap
Initially, flex is used on the header to put the image and
the text side by side for the desktop design. However, doing it in
such a way will make it harder to follow up on the tablet design.
With flex, text will not be able to wrap around the image as depicted
in the tablet design, so flex is now replaced with float in both.